back/requirements.txt
flask
Flask-SQLAlchemy
flask-cors
flask-jwt-extended
bcrypt
python-dotenv
bcrypt
flask-jwt-extended
requests